Fix? socket_audio.scale() function
This function was causing Numpy to raise an error because an array of
type float32 was being saved to an array of type int16. This changes the
result array to float32.

+++ op25/gr-op25_repeater/apps/sockaudio.py
@@ -485,7 +485,7 @@ class socket_audio(object):
 
     def scale(self, data):  # crude amplitude scaler (volume) for S16_LE samples
         arr = np.array(np.frombuffer(data, dtype=np.int16), dtype=np.float32)
-        result = np.zeros(len(arr), dtype=np.int16)
+        result = np.zeros(len(arr), dtype=np.float32)
         arr = np.clip(arr*self.audio_gain, -32767, 32766, out=result)
         return result.tobytes('C')
